cass-prototype/requirements.txt
BeautifulSoup==3.2.1
cassandra-driver==3.3.0
cffi==1.6.0
cql==1.4.0
cryptography==1.3.2
Django==1.9.6
django-allauth==0.18.0
django-appconf==1.0.2
django-cassandra-engine==0.8.1
django-common-helpers==0.7.0
django-compressor==1.4
django-cron==0.4.3
django-debug-toolbar==1.2.1
django-extensions==1.6.6
django-filter==0.11.0
django-pagination==1.0.7
django-rest-swagger==0.3.6
-e git+http://github.com/Tivix/django-userswitch.git@cf3a396269d9f998e1750da3ad7c9a7c60945e77#egg=django_userswitch_fork-dev
djangorestframework==3.3.3
EasyProcess==0.2.2
enum34==1.1.4
Fabric==1.9.1
futures==3.0.5
gnureadline==6.3.3
idna==2.1
ipaddress==1.0.16
ipdb==0.8
ipython==2.2.0
Markdown==2.6.2
model-mommy==1.2.6
numpy==1.11.0
oauthlib==1.1.1
pandas==0.18.1
paramiko==2.0.0
Pillow==2.5.3
pip-tools==0.3.5
psycopg2==2.5.4
pyasn1==0.1.9
pycparser==2.14
Pygments==2.1.3
python-dateutil==2.5.3
python-memcached==1.53
python-openid==2.2.5
pytz==2016.4
PyVirtualDisplay==0.1.5
PyYAML==3.11
requests==2.10.0
requests-oauthlib==0.6.1
selenium==2.48.0
six==1.10.0
sqlparse==0.1.19
thrift==0.9.3
wheel==0.24.0
